Added favicon for most devices
Your site throws an error because it does not have a manifest. I added a manifest and favicon support for moth devices. In a web browser the favicon will show on the browser tab. On Iphone and android it will show as an app icon. You should always add a favicon and a manifest, it helps different devices learn about your app.

https://realfavicongenerator.net/